Added miot features to  Xiaomi Multimode Gateway ZNDMWG03LM
