+++ MiHomeLib/MultimodeGateway/XiaomiGateway3.cs
@@ -3,6 +3,6 @@
 namespace MiHomeLib.MultimodeGateway;
 
 [Obsolete("Please use MultimodeGateway instead, this one will be removed in next releases")]
-public class XiaomiGateway3(string ip, string token, string did = "", int port = 1883) : MultimodeGateway(ip, token, did, port)
+public class XiaomiGateway3(string ip, string token, string did = "", int port = 1883) : MultimodeGatewayBase(ip, token, did, port)
 {
 }
