+++ MiHomeUnitTests/XiaomiMultimodeGatewayTests.cs
@@ -1,495 +1,226 @@
-﻿using FluentAssertions;
-using Moq;
 using Xunit;
-using AutoFixture;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using static MiHomeLib.MultimodeGateway.JsonResponses.ZigbeeHearbeatResponse;
-using static MiHomeLib.MultimodeGateway.JsonResponses.ZigbeeHearbeatResponse.ZigbeeHearbeatItem;
-using static MiHomeLib.MultimodeGateway.JsonResponses.BleAsyncEventResponse;
-using static MiHomeLib.MultimodeGateway.JsonResponses.BleAsyncEventResponse.BleAsyncEventParams;
-using System.Threading.Tasks;
-using MiHomeLib;
+using Moq;
+using FluentAssertions;
 using MiHomeLib.MultimodeGateway;
-using MiHomeLib.MultimodeGateway.JsonResponses;
-using MiHomeLib.MultimodeGateway.Devices;
 using MiHomeUnitTests.MultimodeGatewaySubDevicesTests;
+using System;
 
 namespace MiHomeUnitTests;
 
-public class XiaomiMultimodeGatewayTests: MultimodeGatewayDeviceTests
+public class XiaomiMultimodeGatewayTests : MultimodeGatewayDeviceTests
 {
-    private readonly MultimodeGateway _gateway; 
+    private const string GW_DID = "216264671";
+    private readonly Mock<MultimodeGateway> _gw;
+
     public XiaomiMultimodeGatewayTests()
-    {   
-        _devicesDiscoverer
-            .Setup(x => x.DiscoverZigBeeDevices())
-            .Returns([]);
-
-        _devicesDiscoverer
-            .Setup(x => x.DiscoverBleDevices())
-            .Returns([]);
-
-        _miioTransport
-            .Setup(x => x.SendMessage(It.Is<string>(s => s.Contains("miIO.info"))))
-            .Returns(ToJson(new
-            {
-                id = 1,
-                result = new
-                {
-                    uptime = 502175,
-                    miio_ver = "0.0.9",
-                    mac = "34:EF:44:49:BC:63",
-                    fw_ver = "1.0.5_0008",
-                    hw_ver = "Linux",
-                    ap = new
-                    {
-                        ssid = "ssid1",
-                        bssid = "bssid1",
-                        rssi = -35,
-                        freq = 2437,
-                    },
-                    netif = new
-                    {
-                        localIp = "192.168.1.100",
-                        mask = "255.255.255.0",
-                        gw = "192.168.1.1",
-                    }
-                }
-            }));
-
-        _gateway = new MultimodeGateway(_miioTransport.Object, _mqttTransport.Object, _devicesDiscoverer.Object);
-    }
-    private void SetupZigBeeDevices(List<(string did, string model, int[] props)> devices)
     {
-        foreach (var (did, model, props) in devices)
-        {
-            var getDevicePropResponse = _fixture
-                .Build<GetDevicePropResponse>()
-                .With(x => x.Code, 0)
-                .With(x => x.Result, props)
-                .Without(x => x.Error)
-                .Create();
-
-            _miioTransport
-                .Setup(x => x.SendMessageRepeated(It.Is<string>(s => s.Contains("get_device_prop") && s.Contains(did)), It.IsAny<int>()))
-                .Returns(getDevicePropResponse.ToString());
-        }
-
-        _devicesDiscoverer
-            .Setup(x => x.DiscoverZigBeeDevices())
-            .Returns(devices.Select(x => (x.did, x.model)).ToList());
-    }
-    private void SetupZigBeeMiSpecDevices(List<(string did, string model)> devices)
-    {
-        foreach (var (did, model) in devices)
-        {
-            var getDevicePropResponse = _fixture
-                .Build<GetDevicePropResponse>()
-                .Without(x => x.Code)
-                .Without(x => x.Result)
-                .With(x => x.Error, new GetDevicePropResponse.MiioError { Code = -5015, Message = "device not found" })
-                .Create();
-
-            _miioTransport
-                .Setup(x => x.SendMessageRepeated(It.Is<string>(s => s.Contains("get_device_prop") && s.Contains(did)), It.IsAny<int>()))
-                .Returns(getDevicePropResponse.ToString());
-        }
-
-        _devicesDiscoverer
-            .Setup(x => x.DiscoverZigBeeDevices())
-            .Returns(devices.Select(x => (x.did, x.model)).ToList());
-    }
-    private void SetupBleDevices(List<(string did, int pdid, string mac)> devices)
-    {
-        _devicesDiscoverer
-            .Setup(x => x.DiscoverBleDevices())
-            .Returns(devices);
-    }
-    private void RaiseZigbeeReportEvent(string did, double time, List<(string res, int value)> props)
-    {
-        var zigbeeReport = _fixture
-            .Build<ZigbeeReportResponse>()
-            .With(x => x.Did, did)
-            .With(x => x.Time, time)
-            .Without(x => x.MiSpec)
-            .With(x => x.Params, 
-                props.Select(x => new ZigbeeReportResponse.ZigbeeReportResource
-                {
-                    ResName = x.res,
-                    Value = x.value,
-                }).ToList())
-            .Create();
-
-        _mqttTransport.Raise(x => x.OnMessageReceived += null, "zigbee/send", zigbeeReport.ToString());
-    }
-    private void RaiseZigbeeMiSpecReportEvent(string did, double time, List<(int siid, int piid, object val)> miSpec)
-    {
-        var zigbeeReport = _fixture
-            .Build<ZigbeeReportResponse>()
-            .With(x => x.Did, did)
-            .With(x => x.Time, time)
-            .Without(x => x.Params)
-            .With(x => x.MiSpec, 
-                miSpec.Select(x => new ZigbeeReportResponse.ZigbeeMiSpecItem
-                {
-                    Siid = x.siid,
-                    Piid = x.piid,
-                    Value = x.val,
-                }).ToList())
-            .Create();
-
-        _mqttTransport.Raise(x => x.OnMessageReceived += null, "zigbee/send", zigbeeReport.ToString());
-    }
-    private void RaiseZigbeeHeartbeatEvent(string did, double time, List<(string res, int value)> props)
-    {
-        var @params = new List<ZigbeeHearbeatItem>()
-        {
-            new() {
-                Did = did,
-                Time = time,
-                Zseq = _fixture.Create<byte>(),
-                ResList = props.Select(x => new ZigbeeHearbeatItemResource() { ResName = x.res, Value = x.value }).ToList(),
-            }
-        };
-
-        var zigbeeHeartBeat = _fixture
-            .Build<ZigbeeHearbeatResponse>()
-            .With(x => x.Time, time)
-            .With(x => x.Params, @params)
-            .Create();
-
-        _mqttTransport.Raise(x => x.OnMessageReceived += null, "zigbee/send", zigbeeHeartBeat.ToString());
-    }
-    private void RaiseBleAsyncEvent(int pdid, string did, string mac, double time, List<(int eid, string edata)> data)
-    {
-     var asyncEventResponse = _fixture
-            .Build<BleAsyncEventResponse>()
-            .With(x => x.Params, 
-                _fixture
-                    .Build<BleAsyncEventParams>()
-                    .With(x => x.Dev, 
-                        _fixture
-                            .Build<BleAsyncEventDevice>()
-                            .With(x => x.Did, did)
-                            .With(x => x.Pdid, pdid)
-                            .With(x => x.Mac, mac)
-                            .Create()
-                    )
-                    .With(x => x.Evt, data.Select(x => new BleAsyncEventEvt(){ Eid = x.eid, Edata = x.edata }).ToList())
-                    .With(x => x.Gwts, time)
-                    .Create()
-            )
-            .Create();   
-
-        _mqttTransport.Raise(x => x.OnMessageReceived += null, "miio/report", asyncEventResponse.ToString());
+        _gw = new Mock<MultimodeGateway>(GW_DID, _miioTransport.Object, _mqttTransport.Object, _devicesDiscoverer.Object);
     }
 
     [Fact]
-    public void OnDeviceDiscovered_Works()
+    public void Get_LedEnabled_Should_Return_Correct_Value()
     {
         // Arrange
-        var eventRaised = false;
-        var switchDid = _fixture.Create<string>();
-        var thDid = _fixture.Create<string>();
-        var mac = _fixture.Create<string>()[..12];
-        
-        SetupZigBeeDevices([
-            (switchDid, XiaomiSmartWirelessSwitch.MODEL, [3032, 100, 192, 88]),
-        ]);
+        SetupGetProperties((6, 6, GW_DID, 1));
 
-        SetupBleDevices([
-            (thDid, XiaomiBluetoothHygrothermograph2.PDID, mac),
-        ]);
+        // Act
+        var ledEnabled = _gw.Object.LedEnabled;
 
-        var counter = 0;
-
-        _gateway.OnDeviceDiscoveredAsync += device => 
-        {
-            counter++;
-            eventRaised = true;
-
-            if(device is XiaomiSmartWirelessSwitch sw) sw.Did.Should().Be(switchDid);
-            if(device is XiaomiBluetoothHygrothermograph2 th) th.Did.Should().Be(thDid);
-
-            return Task.CompletedTask;
-        };        
-        
-        // Act    
-        _gateway.DiscoverDevices();
-        
         // Assert
-        eventRaised.Should().BeTrue();
-        counter.Should().Be(2);
+        VerifyGetProperties(6, 6, GW_DID);
+        ledEnabled.Should().BeTrue();
     }
 
     [Fact]
-    public void ZigbeeReportCommand_Works()
+    public void Set_LedEnabled_Should_Work_As_Expected()
     {
         // Arrange
-        var eventRaised = false;
-        var did = _fixture.Create<string>();
-        var time = (double)DateTimeOffset.Now.ToUnixTimeMilliseconds();
+        var expected = 0;
+        SetupSetProperties();
 
-        SetupZigBeeDevices([
-            (did, XiaomiSmartWirelessSwitch.MODEL, [3032, 100, 192, 88]),
-        ]);
-
-        _gateway.OnDeviceDiscoveredAsync += device =>
-        {
-            device.Did.Should().Be(did);
-            var sw = device as XiaomiSmartWirelessSwitch;
-
-            sw.OnClickAsync += clickArgs =>
-            {
-                eventRaised = clickArgs == XiaomiSmartWirelessSwitch.ClickArg.SingleClick;
-                sw.LastTimeMessageReceived.Should().Be(time.UnixMilliSecondsToDateTime());
-                return Task.CompletedTask;
-            };
-
-            return Task.CompletedTask;
-        };
-
-        // Act    
-        _gateway.DiscoverDevices();
-        
-        RaiseZigbeeReportEvent(did, time,[("13.1.85", 1)]);
+        // Act
+        _gw.Object.LedEnabled = false;
 
         // Assert
-        eventRaised.Should().BeTrue();
+        VerifySetProperties(6, 6, GW_DID, expected, 2);
     }
 
     [Fact]
-    public void ZigbeeReportCommandWithMiSpec_Works()
+    public void Get_ArmingMode_Returns_Correct_Mode()
     {
         // Arrange
-        var eventRaised = false;
-        var did = _fixture.Create<string>();
-        var time = (double)DateTimeOffset.Now.ToUnixTimeMilliseconds();
+        var expected = MultimodeGateway.ArmingModeValue.Sleep;
 
-        SetupZigBeeMiSpecDevices([
-            (did, AqaraT1WithNeutral.MODEL),
-        ]);
+        SetupGetProperties((3, 1, GW_DID, expected));
 
-        _gateway.OnDeviceDiscoveredAsync += device =>
-        {
-            device.Did.Should().Be(did);
-            
-            var aqaraRelay = device as AqaraT1WithNeutral;
-
-            aqaraRelay.OnLoadPowerChangeAsync += x =>
-            {
-                eventRaised = true;
-                aqaraRelay.LastTimeMessageReceived.Should().Be(time.UnixMilliSecondsToDateTime());
-                return Task.CompletedTask;
-            };
-
-            return Task.CompletedTask;
-        };
-
-        // Act    
-        _gateway.DiscoverDevices();
-        
-        RaiseZigbeeMiSpecReportEvent(did, time,[(3, 2, 42.2)]);
+        // Act
+        var armingMode = _gw.Object.ArmingMode;
 
         // Assert
-        eventRaised.Should().BeTrue();
+        VerifyGetProperties(3, 1, GW_DID);
+        armingMode.Should().Be(expected);
     }
 
     [Fact]
-    public void ZigbeeHeartbeatCommand_Works()
+    public void Set_ArmingMode_Should_Setup_Correct_Mode()
     {
         // Arrange
-        var voltageEventRaised = false;
-        var batteryEventRaised = false;
-        var oldVoltage = 3032;
-        var newVoltage = 3012;
-        var oldBatteryPerent = 100;
-        var newBatteryPerent = 99;
-        var did = _fixture.Create<string>();
-        var time = (double)DateTimeOffset.Now.ToUnixTimeMilliseconds();
+        var expected = MultimodeGateway.ArmingModeValue.Sleep;
+        SetupSetProperties();
 
-        SetupZigBeeDevices([
-            (did, XiaomiSmartWirelessSwitch.MODEL, [oldVoltage, oldBatteryPerent, 192, 88]),
-        ]);
-
-        _gateway.OnDeviceDiscoveredAsync += device =>
-        {
-            device.Did.Should().Be(did);
-
-            var sw = device as XiaomiSmartWirelessSwitch;
-            
-            sw.OnVoltageChangeAsync += oldValue =>
-            {
-                voltageEventRaised = true;
-                (oldVoltage/1000f).Should().Be(oldValue);
-                sw.Voltage.Should().Be(newVoltage/1000f);
-                sw.LastTimeMessageReceived.Should().Be(time.UnixMilliSecondsToDateTime());
-                return Task.CompletedTask;
-            };
-
-            sw.OnBatteryPercentChange += oldValue =>
-            {
-                batteryEventRaised = true;
-                oldBatteryPerent.Should().Be(oldBatteryPerent);
-                sw.BatteryPercent.Should().Be((byte)newBatteryPerent);
-                sw.LastTimeMessageReceived.Should().Be(time.UnixMilliSecondsToDateTime());
-                return Task.CompletedTask;
-            };
-
-            return Task.CompletedTask;
-        };
-
-        // Act    
-        _gateway.DiscoverDevices();
-        
-        RaiseZigbeeHeartbeatEvent(did, time,[("8.0.2008", newVoltage), ("8.0.2001", newBatteryPerent)]);
+        // Act
+        _gw.Object.ArmingMode = expected;
 
         // Assert
-        voltageEventRaised.Should().BeTrue();
-        batteryEventRaised.Should().BeTrue();
+        VerifySetProperties(3, 1, GW_DID, (int)expected, 2);
     }
 
     [Fact]
-    public void AsyncBleEventMethod_Works()
+    public void Get_AlarmState_Returns_Correct_State()
     {
         // Arrange
-        var temperatureEventRaised = false;
-        var did = _fixture.Create<string>();
-        var mac = _gateway.DecodeMacAddress(_fixture.Create<string>()[..12]);
-        
-        double time = DateTimeOffset.Now.ToUnixTimeSeconds();
+        var expected = MultimodeGateway.AlarmStateValue.NonSecurityAlarm;
 
-        SetupBleDevices([
-            (did, XiaomiBluetoothHygrothermograph2.PDID, mac),
-        ]);
+        SetupGetProperties((3, 22, GW_DID, expected));
 
-        _gateway.OnDeviceDiscoveredAsync += device =>
-        {
-            device.Did.Should().Be(did);
-
-            var th = device as XiaomiBluetoothHygrothermograph2;
-            
-            th.OnTemperatureChangeAsync += oldValue =>
-            {
-                temperatureEventRaised = true;
-                th.LastTimeMessageReceived.Should().Be(time.UnixSecondsToDateTime());
-                return Task.CompletedTask;
-            };
-            
-            return Task.CompletedTask;
-        };
-
-        // Act    
-        _gateway.DiscoverDevices();
-        
-        RaiseBleAsyncEvent(XiaomiBluetoothHygrothermograph2.PDID, did, mac, time, [(4100, "e500")]);
+        // Act
+        var alarmState = _gw.Object.AlarmState;
 
         // Assert
-        temperatureEventRaised.Should().BeTrue();
-    }
-    
-    [Fact]
-    public void GetDevices_When_Any_Returns_DiscoveredDevicesList()
-    {
-        // Arrange
-        var switchDid = _fixture.Create<string>();
-        var thSensorDid = _fixture.Create<string>();
-        var thMonitorDid = _fixture.Create<string>();
-        var mac = _fixture.Create<string>()[..12];
-        
-        SetupZigBeeDevices([
-            (switchDid, XiaomiSmartWirelessSwitch.MODEL, [3032, 100, 192, 88]),
-            (thSensorDid, XiaomiTemperatureHumiditySensor.MODEL, [3032, 100, 192, 88]),
-        ]);
-
-        SetupBleDevices([
-            (thMonitorDid, XiaomiBluetoothHygrothermograph2.PDID, mac),
-        ]);
-
-        // Act    
-        _gateway.DiscoverDevices();
-        var devices = _gateway.GetDevices();
-        
-        // Assert
-        devices.Count.Should().Be(3);
-        devices.Any(x => x.Did == switchDid && x.GetType() == typeof(XiaomiSmartWirelessSwitch)).Should().BeTrue();
-        devices.Any(x => x.Did == thSensorDid && x.GetType() == typeof(XiaomiTemperatureHumiditySensor)).Should().BeTrue();
-        devices.Any(x => x.Did == thMonitorDid && x.GetType() == typeof(XiaomiBluetoothHygrothermograph2)).Should().BeTrue();
-    }
-    
-    [Fact]
-    public void GetDevices_When_NoOne_Returns_EmptyList()
-    {
-        // Arrange
-
-        // Act    
-        _gateway.DiscoverDevices();
-        
-        // Assert
-        _gateway.GetDevices().Count.Should().Be(0);
+        VerifyGetProperties(3, 22, GW_DID);
+        alarmState.Should().Be(expected);
     }
 
     [Fact]
-    public void GetDeviceByDid_When_NotFound_Returns_Null()
+    public void Set_AlarmState_Should_Setup_Correct_State()
     {
         // Arrange
-        var did = _fixture.Create<string>();
+        var expected = MultimodeGateway.AlarmStateValue.NoAlarm;
+        SetupSetProperties();
 
-        // Act    
-        _gateway.DiscoverDevices();
-        
+        // Act
+        _gw.Object.AlarmState = expected;
+
         // Assert
-        _gateway.GetDeviceByDid<XiaomiSmartWirelessSwitch>(did).Should().BeNull();
+        VerifySetProperties(3, 22, GW_DID, (int)expected, 2);
     }
 
     [Fact]
-    public void GetDeviceByDid_WhenExistsAndHasCorrectType_Returns_Device()
+    public void Get_DelayForArmingMode_Returns_Correct_Delay()
     {
         // Arrange
-        var switchDid = _fixture.Create<string>();
-        
-        SetupZigBeeDevices([
-            (switchDid, XiaomiSmartWirelessSwitch.MODEL, [3032, 100, 192, 88]),
-            (_fixture.Create<string>(), XiaomiTemperatureHumiditySensor.MODEL, [3032, 100, 192, 88]),
-        ]);
+        byte expectedDelay = 5;
 
-        SetupBleDevices([
-            (_fixture.Create<string>(), XiaomiBluetoothHygrothermograph2.PDID, _fixture.Create<string>()),
-        ]);
+        SetupGetProperties((3, 18, GW_DID, expectedDelay));
 
-        // Act    
-        _gateway.DiscoverDevices();
-        
+        // Act
+        var delay = _gw.Object.GetDelayTimeForArmingMode(MultimodeGateway.ArmingModeValue.Sleep);
+
         // Assert
-        _gateway
-            .GetDeviceByDid<XiaomiSmartWirelessSwitch>(switchDid)
-                .Should()
-                .NotBeNull()
-                .And
-                .Match<XiaomiSmartWirelessSwitch>(x => x.Did == switchDid);
+        VerifyGetProperties(3, 18, GW_DID);
+        delay.Should().Be(expectedDelay);
     }
 
     [Fact]
-    public void GetDeviceByDid_WhenHasWrongType_Returns_Null()
+    public void Set_DelayForArmingMode_Should_Work_As_Expected()
     {
         // Arrange
-        var switchDid = _fixture.Create<string>();
-        
-        SetupZigBeeDevices([
-            (switchDid, XiaomiSmartWirelessSwitch.MODEL, [3032, 100, 192, 88]),
-            (_fixture.Create<string>(), XiaomiTemperatureHumiditySensor.MODEL, [3032, 100, 192, 88]),
-        ]);
+        byte expected = 50;
+        SetupSetProperties();
 
-        SetupBleDevices([
-            (_fixture.Create<string>(), XiaomiBluetoothHygrothermograph2.PDID, _fixture.Create<string>()),
-        ]);
+        // Act
+        _gw.Object.SetDelayTimeForArmingMode(MultimodeGateway.ArmingModeValue.Sleep, expected);
 
-        // Act    
-        _gateway.DiscoverDevices();
-        
         // Assert
-        _gateway.GetDeviceByDid<XiaomiTemperatureHumiditySensor>(switchDid).Should().BeNull();
-    }   
+        VerifySetProperties(3, 18, GW_DID, expected, 2);
+    }
+
+    [Fact]
+    public void Set_Incorrect_DelayForArmingMode_Should_Throw_Exception()
+    {
+        // Arrange
+        byte expected = 65;
+        SetupSetProperties();
+
+        // Act
+        var result = _gw.Invoking(x => x.Object.SetDelayTimeForArmingMode(MultimodeGateway.ArmingModeValue.Sleep, expected));
+
+        // Assert
+        result.Should().Throw<ArgumentOutOfRangeException>();
+    }
+
+    [Fact]
+    public void Get_AlarmDurationForArmingMode_Returns_Correct_Duration()
+    {
+        // Arrange
+        uint expected = 1_000_000;
+
+        SetupGetProperties((3, 19, GW_DID, expected));
+
+        // Act
+        var duration = _gw.Object.GetAlarmDurationForArmingMode(MultimodeGateway.ArmingModeValue.Sleep);
+
+        // Assert
+        VerifyGetProperties(3, 19, GW_DID);
+        duration.Should().Be(expected);
+    }
+
+    [Fact]
+    public void Set_AlarmDurationForArmingMode_Should_Work_As_Expected()
+    {
+        // Arrange
+        uint expected = 2_000_000;
+        SetupSetProperties();
+
+        // Act
+        _gw.Object.SetAlarmDurationForArmingMode(MultimodeGateway.ArmingModeValue.Sleep, expected);
+
+        // Assert
+        VerifySetProperties(3, 19, GW_DID, expected, 2);
+    }
+
+    [Fact]
+    public void Set_Incorrect_AlarmDurationForArmingMode_Should_Throw_Exception()
+    {
+        // Arrange
+        uint expected = 3_000_000_000;
+        SetupSetProperties();
+
+        // Act
+        var result = _gw.Invoking(x => x.Object.SetAlarmDurationForArmingMode(MultimodeGateway.ArmingModeValue.Sleep, expected));
+
+        // Assert
+        result
+            .Should()
+            .Throw<ArgumentOutOfRangeException>();
+    }
+
+    [Fact]
+    public void Get_AlarmVolumenForArmingMode_Returns_Correct_VolumeLevel()
+    {
+        // Arrange
+        var expected = MultimodeGateway.VolumeLeveValue.Middle;
+        SetupGetProperties((3, 20, GW_DID, expected));
+
+        // Act
+        var volumeLevel = _gw.Object.GetAlarmVolumeForArmingMode(MultimodeGateway.ArmingModeValue.Sleep);
+
+        // Assert
+        VerifyGetProperties(3, 20, GW_DID);
+        volumeLevel.Should().Be(expected);
+    }
+
+    [Fact]
+    public void Set_AlarmVolumenForArmingMode_Should_Work_As_Expected()
+    {
+        // Arrange
+        var expected = MultimodeGateway.VolumeLeveValue.High;
+        SetupSetProperties();
+
+        // Act
+        _gw.Object.SetAlarmVolumeForArmingMode(MultimodeGateway.ArmingModeValue.Sleep, expected);
+
+        // Assert
+        VerifySetProperties(3, 20, GW_DID, expected, 2);
+    }
 }
